% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bma.R
\name{bma_ens_models}
\alias{bma_ens_models}
\title{Fit BMA coefficients for beta distributions with discrete component at 1 for
each ensemble member Data should be pre-processed and normalized to [0,1] so
that clipped values are exactly 1}
\usage{
bma_ens_models(tel, ens, bma_distribution = "beta", max_power = NA,
  lr_formula = y ~ x, A_transform = NA, lm_formula = y ~ x + 0,
  B_transform = NA, percent_clipping_threshold = 0.995, tol = 0.001,
  ...)
}
\arguments{
\item{tel}{Vector of training telemetry data on [0,1] OR a matrix [time x
member] of training telemetry data tailored to each member}

\item{ens}{Matrix of training ensemble member data [time x member] on [0,1]}

\item{bma_distribution}{Either "beta" or "truncnorm" to select type of
distribution for member kernel dressing}

\item{lr_formula}{Formula in terms of x,y for logistic regression model,
defaults to "y ~ x". Requires a negative x intercept to model PoC < 0.5.}

\item{A_transform}{A function for transforming forecast data before logistic
regression to get a's (optional)}

\item{lm_formula}{Formula in terms of x,y for linear regression model,
defaults to "y ~ x + 0"}

\item{B_transform}{A function for transforming forecast data before linear
regression to get b's (optional)}

\item{percent_clipping_threshold}{[0,1] Power is designated as clipped when
above this percentage of the maximum power}

\item{tol}{A tolerance for determining if normalized values all are <=1
(defaults to 0.001)}

\item{...}{Optional arguments to pass to EM algorithm}
}
\value{
A list for a discrete-continuous mixture model with beta distribution
}
\description{
Fit BMA coefficients for beta distributions with discrete component at 1 for
each ensemble member Data should be pre-processed and normalized to [0,1] so
that clipped values are exactly 1
}
