% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ts_forecast.R
\name{Brier}
\alias{Brier}
\title{Get Brier score at power thresholds}
\usage{
Brier(ts, tel, thresholds)
}
\arguments{
\item{ts}{A ts_forecast object}

\item{tel}{A list of the telemetry values}

\item{thresholds}{Thresholds in units of the forecast. Can be a scalar or a
vector}
}
\value{
the Brier score
}
\description{
Get Brier score at power thresholds
}
