% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ts_forecast.R
\name{Brier_quantile}
\alias{Brier_quantile}
\title{Get Brier score at a certain probability of exceedance}
\usage{
Brier_quantile(ts, tel, PoE, ...)
}
\arguments{
\item{ts}{A ts_forecast object}

\item{tel}{A list of the telemetry values}

\item{PoE}{Threshold probability of exceedance, numeric [0,1]}

\item{...}{optional arguments to equalize_telemetry_forecast_length}
}
\value{
the Brier score
}
\description{
This is calculated
with a constant probability threshold, rather than a constant value threshold
}
