% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prob_forecast.R
\name{calc_cvar}
\alias{calc_cvar}
\title{Calculate VaR and CVaR from sampled data. CVaR calculation is done directly from the samples, rather than estimated from a fitted distribution.}
\usage{
calc_cvar(samples, epsilon = c(0.05, 0.95))
}
\arguments{
\item{samples}{Numeric vector}

\item{epsilon}{Probability levels for lower/upper tail VaR/CVaR calculations, defaults to c(0.05, 0.95)}
}
\value{
list of var, cvar
}
\description{
Calculate VaR and CVaR from sampled data. CVaR calculation is done directly from the samples, rather than estimated from a fitted distribution.
}
