% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ts_forecast.R
\name{calc_forecasts}
\alias{calc_forecasts}
\title{Calculate a time series list of power forecasts.}
\usage{
calc_forecasts(x, sun_up, start_time, time_step, scale, location, method,
  n, epsilon)
}
\arguments{
\item{x}{A list of training data the length of the time-series. Each element should be [ntrain x 1] matrix of training data (for scale=='site')
or a [ntrain x nsites] matrix for scale=='region' or 'total'}

\item{sun_up}{Logical vector the length of the time-series}

\item{start_time}{A lubridate time stamp}

\item{time_step}{Time step in hours}

\item{scale}{One of 'site', 'region', 'total'}

\item{location}{A string}

\item{method}{One of 'gaussian', 'empirical', 'vine'}

\item{n}{An integer: Number of samples to take}

\item{epsilon}{Probability levels for lower/upper tail VaR/CVaR calculations, defaults to c(0.05, 0.95)}
}
\value{
A list of forecasts. Forecast is NA for times when sun is down.
}
\description{
Calculate a time series list of power forecasts.
}
