% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ts_forecast.R
\name{calc_PIT_histogram}
\alias{calc_PIT_histogram}
\title{Calculate data for probability integral transform histogram}
\usage{
calc_PIT_histogram(ts, tel, nbins, ...)
}
\arguments{
\item{ts}{A ts_forecast object}

\item{tel}{A list of the telemetry values}

\item{nbins}{To specify number of bars in the histogram}

\item{...}{optional arguments to equalize_telemetry_forecast_length}
}
\value{
List of the bin locations, their relative frequency, and the bin
  width
}
\description{
Calculate data for probability integral transform histogram
}
