% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prob_forecast.R
\name{calc_quantiles.fc_emos}
\alias{calc_quantiles.fc_emos}
\title{Calculate forecast quantiles}
\usage{
\method{calc_quantiles}{fc_emos}(x, model, quantiles = seq(0.001, 0.999,
  by = 0.001), ...)
}
\arguments{
\item{x}{fc_emos object}

\item{quantiles}{Sequence of quantiles in (0,1)}
}
\value{
A list of q, the quantiles on [0, 1], and x, the estimated values
}
\description{
Calculate forecast quantiles
}
