% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prob_forecast.R
\name{calc_quantiles.fc_vine}
\alias{calc_quantiles.fc_vine}
\title{Calculate forecast quantiles from samples of the vine copula}
\usage{
\method{calc_quantiles}{fc_vine}(x, samples = NA, samples.u = NA,
  quantiles = seq(0.001, 0.999, by = 0.001))
}
\arguments{
\item{x}{fc_vine object}

\item{samples}{(optional) previously obtained 1-D samples to use instead of new sampling, e.g. for coordination with cVaR calculation}

\item{samples.u}{(optional) previously obtained n-d samples to use instead of new sampling, e.g. for static vine copula model.
samples overrides samples.u if both are given}

\item{quantiles}{Sequence of quantiles in (0,1)}
}
\value{
A list of q, the quantiles on [0, 1], and x, the estimated values
}
\description{
Calculate forecast quantiles from samples of the vine copula
}
