% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prob_forecast.R
\name{calc_quantiles}
\alias{calc_quantiles}
\title{Register generic quantiles function}
\usage{
calc_quantiles(x, ...)
}
\arguments{
\item{x}{A prob_forecast object}
}
\description{
Register generic quantiles function
}
