% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prob_forecast.R
\name{calc_transforms}
\alias{calc_transforms}
\title{Calculate lists of variable-to-uniform domain transforms for all dimensions}
\usage{
calc_transforms(dat, d, training_transform_type, results_transform_type,
  ...)
}
\arguments{
\item{dat}{training data matrix}

\item{d}{Number of dimensions}

\item{training_transform_type}{Transform of training data into uniform domain (see marg_transform "cdf.method")}

\item{results_transform_type}{Transform of copula results back into variable domain (see marg_transform "cdf.method")}

\item{...}{Optional arguments to marg_transform}
}
\value{
list of "training" and "results" transforms to use.
}
\description{
Calculate lists of variable-to-uniform domain transforms for all dimensions
}
