% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prob_forecast.R
\name{cdf_subfunction}
\alias{cdf_subfunction}
\title{Get cumulative distribution for individual ensemble member}
\usage{
cdf_subfunction(param1, param2, poc, w, xseq, i.thresh, bma_distribution,
  max_power)
}
\arguments{
\item{param1}{First parameter: alpha for a beta distribution; mean for a truncnorm}

\item{param2}{First parameter: beta for a beta distribution; sd for a truncnorm}

\item{bma_distribution}{"beta" or "truncnorm"}
}
\description{
Get cumulative distribution for individual ensemble member
}
