% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prob_forecast.R
\name{CVAR.fc_kde}
\alias{CVAR.fc_kde}
\title{Calculate VaR and CVaR by trapezoidal integration.}
\usage{
\method{CVAR}{fc_kde}(x, epsilon = c(0.05, 0.95))
}
\arguments{
\item{x}{fc_kde object}

\item{epsilon}{Probability levels for lower/upper tail VaR/CVaR calculations, defaults to c(0.05, 0.95)}
}
\value{
list of var, cvar
}
\description{
Calculate VaR and CVaR by trapezoidal integration.
}
