% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bma.R
\name{dbeta_gamma_rho}
\alias{dbeta_gamma_rho}
\title{Get density of beta distribution with gamma, rho parameterization at given
value}
\usage{
dbeta_gamma_rho(x, g, rho)
}
\arguments{
\item{x}{Value or vector of values [0,1]}

\item{g}{Gamma = alpha + beta}

\item{rho}{rho = alpha/(alpha + beta)}
}
\description{
Get density of beta distribution with gamma, rho parameterization at given
value
}
