% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analogs.R
\name{delle_monache_distance}
\alias{delle_monache_distance}
\title{Calculate Delle Monache distance metric for the physical features for a potential analog}
\usage{
delle_monache_distance(h, f, weights, sigmas)
}
\arguments{
\item{h}{a potential analog, [time along matching window x physical feature]}

\item{f}{forecast of features, [time along matching window x physical feature]}

\item{weights}{Vector of weights to use for each feature}

\item{sigmas}{Standard deviations of the features over the historical record}
}
\value{
Delle Monache distance
}
\description{
Calculate Delle Monache distance metric for the physical features for a potential analog
}
