% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dhhh4sims.R
\name{dhhh4sims}
\alias{dhhh4sims}
\title{\code{hhh4}-Based Forecast Distributions}
\usage{
dhhh4sims(sims, model)
}
\arguments{
\item{sims}{a \code{"hhh4sims"} object from
\code{\link[surveillance]{simulate.hhh4}}.}

\item{model}{the \code{"\link[surveillance]{hhh4}"} object underlying
\code{sims}.}
}
\value{
a \code{function(x, tp = 1, log = FALSE)}, which takes a
    vector of \code{model$nUnit} counts and calculates the
    (\code{log}-)probability of observing these counts (given the
    \code{model}) at the \code{tp}'th time point of the simulation
    period.
}
\description{
The function \code{dhhh4sims} constructs a (non-vectorized)
probability mass function from the result of
\code{\link[surveillance]{simulate.hhh4}} (and the corresponding
model), as a function of the time point within the simulation period.
The distribution at each time point is obtained as a mixture of
negative binomial (or Poisson) distributions based on the samples from
the previous time point.
}
\references{
\Sexpr[stage=build,results=rd]{tools::toRd(citation("HIDDA.forecasting"))}
}
\seealso{
\code{\link{logs_hhh4sims}} where this function is used.
}
\author{
Sebastian Meyer
}
\keyword{distribution}
