% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scores.R
\name{dss_lnorm}
\alias{dss_lnorm}
\title{Dawid-Sebastiani Score (DSS) for Log-Normal Predictions}
\usage{
dss_lnorm(x, meanlog, sdlog)
}
\arguments{
\item{x}{the observed counts.}

\item{meanlog, sdlog}{parameters of the log-normal distribution, i.e., mean
and standard deviation of the distribution on the log scale.}
}
\value{
scores for the predictions of the observations in \code{x} (maintaining
their dimensions).
}
\description{
Computes the DSS for the log-normal distribution similar to
\code{\link[surveillance:dss]{surveillance::dss()}} for the Poisson and negative binomial distribution.
Predictions are assumed to have a \link[=stats:plnorm]{LN}(\code{meanlog}, \code{sdlog})
distribution. The function is vectorized and preserves the dimensions of the
input.
}
\author{
Sebastian Meyer
}
