% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/emos.R
\name{emos_model}
\alias{emos_model}
\title{Fit EMOS coefficients a, b_k, c, d for truncated normal b_k, c, and d are
parameterized as squares to control non-negativity of regression and scale
parameters, like the "square" coefRule and varRule in the ensembleMOS
package.}
\usage{
emos_model(tel, ens, max_power, par_init = NA)
}
\arguments{
\item{tel}{Vector of training telemetry data}

\item{ens}{Matrix of training ensemble member data [time x member]}

\item{max_power}{site AC rating, for upper limit of truncated normal}

\item{par_init}{Initial parameter values as a list with a, b, c, d}
}
\description{
Fit EMOS coefficients a, b_k, c, d for truncated normal b_k, c, and d are
parameterized as squares to control non-negativity of regression and scale
parameters, like the "square" coefRule and varRule in the ensembleMOS
package.
}
