% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ts_forecast.R
\name{equalize_telemetry_forecast_length}
\alias{equalize_telemetry_forecast_length}
\title{Equalize lengths of telemetry and forecast vectors}
\usage{
equalize_telemetry_forecast_length(tel, fc, agg = TRUE,
  align = "end-of-hour")
}
\arguments{
\item{tel}{A vector of the telemetry values OR a single value of the length
of the telemetry values}

\item{fc}{A vector of data from the time series forecast OR a single value of
the length of the forecast values}

\item{agg}{Boolean, TRUE to aggregate telemetry to forecast resolution, FALSE
to expand forecast to telemetry resolution}

\item{align}{Can be "end-of-hour", "half-hour " NaN first, telemetry lags
Defaults to "end-of-hour". Currently half hour approaches expand
forecast the same way.}
}
\value{
list of the telemetry and forecast data vectors, of equal length
}
\description{
Preprocess for metrics evaluations, for when telemetry is at finer time
resolution than the forecast. The forecast and telemetry can be at different
time resolutions, so long as telemetry is a multiple of the forecast.
}
