% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prob_forecast.R
\name{error_check_calc_quantiles_input}
\alias{error_check_calc_quantiles_input}
\title{Error check quantiles are in (0,1)}
\usage{
error_check_calc_quantiles_input(quantiles)
}
\arguments{
\item{quantiles}{A vector}
}
\description{
Error check quantiles are in (0,1)
}
