% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ts_forecast.R
\name{eval_avg_crps}
\alias{eval_avg_crps}
\title{Get the average estimated CRPS (continuous ranked probability score) for the forecast;
the score at each time point is estimated from sampled data.}
\usage{
eval_avg_crps(ts, actuals)
}
\arguments{
\item{ts}{A ts_forecast object}

\item{actuals}{A list of the realized values}
}
\description{
Get the average estimated CRPS (continuous ranked probability score) for the forecast;
the score at each time point is estimated from sampled data.
}
