% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/unfinished code bits.R
\name{eval_mer}
\alias{eval_mer}
\title{Get Missing Rate Error. Indicates underdispersive (+) or overdispersive (-) forecast
This is different from reliability index, right? (can't remember what that is...)
SIMILAR THOUGHT ON ENSEMBLE APPLICATION AS RI ABOVE.}
\usage{
eval_mer(ts, actuals)
}
\arguments{
\item{ts}{A ts_forecast object}

\item{actuals}{A list of the realized values}
}
\value{
numeric, missing rate error
}
\description{
Get Missing Rate Error. Indicates underdispersive (+) or overdispersive (-) forecast
This is different from reliability index, right? (can't remember what that is...)
SIMILAR THOUGHT ON ENSEMBLE APPLICATION AS RI ABOVE.
}
