% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prob_forecast.R
\name{fc_binned}
\alias{fc_binned}
\title{A binned probability forecast}
\usage{
fc_binned(data.input, location, time, max_power, ...)
}
\arguments{
\item{data.input}{A numeric vector of ensemble members}

\item{location}{A string}

\item{time}{A lubridate time stamp}

\item{max_power}{Maximum power for normalizing forecast to [0,1]}
}
\value{
A probabilistic forecast object
}
\description{
Initialize a univariate probabilistic power forecast for a specific time
point by ranking ensemble members and *linearly interpolating*. For a classic
stepped forecast, see \code{\link{fc_empirical}}.Its rank quantiles are the
basic quantiles estimated from the ensemble members; the quantiles are the
rank quantiles iterpolated to the quantiles of interest (i.e., 10%, 20%,
etc...).
}
\seealso{
Other forecast classes: \code{\link{fc_bma}},
  \code{\link{fc_emos}}, \code{\link{fc_empirical}}
}
\concept{forecast classes}
