% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prob_forecast.R
\name{fc_emos}
\alias{fc_emos}
\title{An EMOS forecast}
\usage{
fc_emos(data.input, location, time, model, max_power, ...)
}
\arguments{
\item{data.input}{A vector of ensemble members}

\item{location}{A string}

\item{time}{A lubridate time stamp}

\item{model}{A pre-fit BMA model from bma_ens_models; list of a, b, c, d
parameters}

\item{max_power}{Maximum power for normalizing forecast to [0,1]}

\item{...}{Additional parameters}
}
\value{
A probabilistic forecast object
}
\description{
Initialize a univariate probabilistic power forecast for a specific time
point using ensemble model output statistics (EMOS)
}
\seealso{
Other forecast classes: \code{\link{fc_binned}},
  \code{\link{fc_bma}}, \code{\link{fc_empirical}}
}
\concept{forecast classes}
