% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prob_forecast.R
\name{fc_empirical}
\alias{fc_empirical}
\title{A forecast from an empirical CDF}
\usage{
fc_empirical(data.input, location, time, ...)
}
\arguments{
\item{data.input}{A numeric vector of training data}

\item{location}{A string}

\item{time}{A lubridate time stamp}
}
\value{
A probabilistic forecast object
}
\description{
Initialize a forecast using an empirical (stepped, not interpolated) CDF of a
training set. Can be applied for a climatology, raw ensemble, or persistence
ensemble type forecast.
}
\seealso{
Other forecast classes: \code{\link{fc_binned}},
  \code{\link{fc_bma}}, \code{\link{fc_emos}}
}
\concept{forecast classes}
