% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prob_forecast.R
\name{fc_vine}
\alias{fc_vine}
\title{Initialize a probabilistic power forecast for a specific time point, using an n-dimensional vine copula.
Assumes training data already captures differences in magnitude (i.e., power rating) amongst sites.
The "copula" field of the data.input list can be a matrix of training data [ntrain x nsites] OR a pre-trained vinecop model
The "marginals" field of the data.input list can be a matrix of training data [ntrain x nsites] OR a list of single-site prob_forecast objects}
\usage{
fc_vine(data.input, location, time,
  training_transform_type = "empirical",
  results_transform_type = "empirical", n = 10000, samples.u = NA,
  ...)
}
\arguments{
\item{data.input}{A list of "copula" and "marginals" inputs}

\item{location}{A string}

\item{time}{A lubridate time stamp}

\item{training_transform_type}{Transform of training data into uniform domain (see marg_transform "cdf.method")}

\item{results_transform_type}{Transform of copula results back into variable domain (see marg_transform "cdf.method")}

\item{n}{An integer, number of copula samples to take}

\item{samples.u}{(optional) A precalculated set of n-dimensional CDF samples from rvinecop}

\item{...}{optional arguments to the marginal estimator}
}
\value{
An n-dimensional probabilistic forecast object from vine copulas
}
\description{
Initialize a probabilistic power forecast for a specific time point, using an n-dimensional vine copula.
Assumes training data already captures differences in magnitude (i.e., power rating) amongst sites.
The "copula" field of the data.input list can be a matrix of training data [ntrain x nsites] OR a pre-trained vinecop model
The "marginals" field of the data.input list can be a matrix of training data [ntrain x nsites] OR a list of single-site prob_forecast objects
}
