% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analogs.R
\name{feature_distance}
\alias{feature_distance}
\title{Calculate the distance metric for one physical feature for a potential analog
Raw distance is 0 if BOTH forecast and historical are missing
to accommodate windows that overlap sunrise/set. Distance is still NA if only one is missing.
Matrix slicing is done internally to handle half_interval of 0 or matrix of single feature (to avoid vectors)}
\usage{
feature_distance(weight, sigma, f, h)
}
\arguments{
\item{weight}{weight [0,1] assigned to this feature}

\item{sigma}{Standard deviation of the feature over the historical record}

\item{f}{A vector forecast of one features}

\item{h}{An potential analog forecast of one feature}
}
\value{
feature distance
}
\description{
Calculate the distance metric for one physical feature for a potential analog
Raw distance is 0 if BOTH forecast and historical are missing
to accommodate windows that overlap sunrise/set. Distance is still NA if only one is missing.
Matrix slicing is done internally to handle half_interval of 0 or matrix of single feature (to avoid vectors)
}
