% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/marg_transform.R
\name{from_uniform}
\alias{from_uniform}
\title{Inverse transform from uniform to variable domain}
\usage{
from_uniform(c, u)
}
\arguments{
\item{c}{A marg_transform object}

\item{u}{A vector of evaluation points to transform}
}
\value{
A vector linearly interpolated from the uniform to variable domain
}
\description{
Inverse transform from uniform to variable domain
}
