% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prob_forecast.R
\name{get_1d_samples.fc_vine}
\alias{get_1d_samples.fc_vine}
\title{Sample the vine copula model and sum to calculate samples of the univariate, aggregate power forecast}
\usage{
\method{get_1d_samples}{fc_vine}(x, samples.u = NA)
}
\value{
A column matrix of aggregate powers
}
\description{
Sample the vine copula model and sum to calculate samples of the univariate, aggregate power forecast
}
