% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prob_forecast.R
\name{get_beta_distribution_geometry_code}
\alias{get_beta_distribution_geometry_code}
\title{0 -> Missing
1 -> U-type
2 -> Reverse J-type
3 -> J-type
4 -> Upside-down U type}
\usage{
get_beta_distribution_geometry_code(bma_distribution, alpha, beta)
}
\arguments{
\item{bma_distribution}{"beta" or "truncnorm"; "truncnorm" returns 0}

\item{alpha}{Value of beta distribution's alpha parameter}

\item{beta}{Value of beta distribution's beta parameter}
}
\description{
0 -> Missing
1 -> U-type
2 -> Reverse J-type
3 -> J-type
4 -> Upside-down U type
}
