% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prob_forecast.R
\name{get_discrete_continuous_model}
\alias{get_discrete_continuous_model}
\title{Get the discrete and continuous components of the weighted mixture model as
well as the member-specific components}
\usage{
get_discrete_continuous_model(x, xseq = seq(0, 1, 1e-04), discrete = F)
}
\arguments{
\item{x}{fc_bma object}

\item{xseq}{Vector of x values in [0,1] to evaluate}

\item{discrete}{Boolean of whether to return density with a discrete
component (approximate) or with the continuous estimation (congruent with
CDF)}
}
\value{
A list of the discrete components (PoC) and continuous density (pdf)
  and distribution (cdf) for each member
}
\description{
Get the discrete and continuous components of the weighted mixture model as
well as the member-specific components
}
