% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ts_forecast.R
\name{get_forecast_class}
\alias{get_forecast_class}
\title{Look-up function of the forecast class type}
\usage{
get_forecast_class(scale, method)
}
\arguments{
\item{scale}{One of 'site', 'region', 'total'}

\item{method}{One of 'binned' if scale is 'site', else one of gaussian',
'empirical', 'vine'}
}
\value{
A function to initialize a forecast of the desired type
}
\description{
Look-up function of the forecast class type
}
