% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bma.R
\name{get_gamma}
\alias{get_gamma}
\title{Get gamma parameter of beta distribution for single instance}
\usage{
get_gamma(mu, sigma)
}
\arguments{
\item{mu}{beta mean value for this forecast (AKA rho)}

\item{sigmas}{standard deviation}
}
\description{
Get gamma parameter of beta distribution for single instance
}
