% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analogs.R
\name{get_historical_analogs}
\alias{get_historical_analogs}
\title{Assumes that the forecast data and historical analogs are at the same time-resolution
To calculate standard deviation of each feature, zero values are skipped.}
\usage{
get_historical_analogs(f_test, h_train, h_real, n, weights,
  sigmas = FALSE)
}
\arguments{
\item{f_test}{matrix of the forecast data to fit on [time along matching window x physical feature]}

\item{h_train}{array of historical forecast data [potential analog time x time along matching window x physical feature]. Matching window time can be a singleton dimension}

\item{h_real}{Historical realized value of interest, e.g., power (equivalent to a kNN classification)}

\item{n}{Integer, number of historical analogs to pick}

\item{weights}{Vector of weights to use for each feature}

\item{sigmas}{(optional) Vector of standard deviations to use for each feature. Can be re-calculated if feature has redundancies to make matrix structure.}
}
\value{
A list of the analogs, including observed value, the forecast along the matching window, and the distance metric.
}
\description{
Assumes that the forecast data and historical analogs are at the same time-resolution
To calculate standard deviation of each feature, zero values are skipped.
}
