% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prob_forecast.R
\name{get_joint_density_grid.fc_vine}
\alias{get_joint_density_grid.fc_vine}
\title{Estimate the joint probability density}
\usage{
\method{get_joint_density_grid}{fc_vine}(x, k = 100)
}
\arguments{
\item{x}{A fc_vine object}

\item{k}{Integer or vector of number of samples to take along each dimension. If given an integer, all dimensions have the same number of samples.}
}
\value{
an joint probability density array
}
\description{
Estimate the joint probability density
}
