% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prob_forecast.R
\name{get_joint_density_grid}
\alias{get_joint_density_grid}
\title{Register generic joint density function}
\usage{
get_joint_density_grid(x, ...)
}
\arguments{
\item{x}{An n-dimension prob_forecast object}
}
\description{
Register generic joint density function
}
