% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ts_forecast.R
\name{get_metric_time_series}
\alias{get_metric_time_series}
\title{Get a time-series of the forecast performance in terms of one metric}
\usage{
get_metric_time_series(metric, ts, tel, normalize.by, metricArgs = NULL,
  ...)
}
\arguments{
\item{metric}{A function for prob_forecast objects}

\item{ts}{A ts_forecast object}

\item{tel}{Vector of telemetry}

\item{normalize.by}{A vector or scalar}

\item{metricArgs}{(optional) a list of additional arguments to the metrics
function}

\item{...}{optional arguments to equalize_telemetry_forecast_length}
}
\value{
A vector
}
\description{
Get a time-series of the forecast performance in terms of one metric
}
