% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility.R
\name{get_netcdf_forecast_data}
\alias{get_netcdf_forecast_data}
\title{Load data from a NETCDF file of ensemble forecasts Assumed file dimensions:
Day x Hour x Site x Lead time x Ensemble member Returns an array of data:
[site x member x time] or [site x member x time x lead time] Site selection
and member selection can all be vectors of non-consecutive values Time-point
selection is a consecutive sequence}
\usage{
get_netcdf_forecast_data(fname, members, sites, lead_times, date_start,
  date_end, date_data_start = lubridate::ymd(20160101),
  ts_per_day = 24, vname = "power", truncate = F,
  site_max_power = NA)
}
\arguments{
\item{fname}{file name}

\item{members}{A vector of member indices}

\item{sites}{A vector of sites}

\item{lead_times}{Forecast lead time or a set of lead times}

\item{date_start}{A lubridate: Start date of data to load}

\item{date_end}{A lubridate: End date of data to load}

\item{date_data_start}{A lubridate: Date of first day in file}

\item{truncate}{Boolean: Whether or not to truncate the forecasts at the site
maximum power}

\item{site_max_power}{A vector of the maximum power at ALL sites (not just
those listed in sites)}
}
\description{
Load data from a NETCDF file of ensemble forecasts Assumed file dimensions:
Day x Hour x Site x Lead time x Ensemble member Returns an array of data:
[site x member x time] or [site x member x time x lead time] Site selection
and member selection can all be vectors of non-consecutive values Time-point
selection is a consecutive sequence
}
