% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility.R
\name{get_netcdf_telemetry_data}
\alias{get_netcdf_telemetry_data}
\title{Load data from a NETCDF file of telemetry
Assumed file dimensions: Day x Hour x Site
Returns an array of data: [site x time]
Site selection can be vector of non-consecutive values
Time-point selection is a consecutive sequence}
\usage{
get_netcdf_telemetry_data(fname, sites, date_start, date_end,
  date_data_start = lubridate::ymd(20160101), ts_per_day = 288,
  vname = "hsl_power")
}
\arguments{
\item{fname}{file name}

\item{sites}{A vector of sites}

\item{date_start}{A lubridate: Start date of data to load}

\item{date_end}{A lubridate: End date of data to load}
}
\description{
Load data from a NETCDF file of telemetry
Assumed file dimensions: Day x Hour x Site
Returns an array of data: [site x time]
Site selection can be vector of non-consecutive values
Time-point selection is a consecutive sequence
}
