% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bma.R
\name{get_poc}
\alias{get_poc}
\title{Get PoC (probability of clipping) for single instance PoC is estimated from
logit(PoC)=a0 + a1*f + a2*1[obs==1] This seems inverted from the PoP
defintion from Will Kleiber's code, which reads to me like it should be 1-PoP}
\usage{
get_poc(FCST, A0, A1, A2, A_transform = NA)
}
\arguments{
\item{FCST}{single forecast}

\item{A0}{intercept of logistic regression of PoC}

\item{A1}{slope of logistic regression of PoC}

\item{A2}{coefficient of clipped forecast indicator [FCST==1] in logistic
regression of PoC}

\item{OBS}{single observation}
}
\description{
Get PoC (probability of clipping) for single instance PoC is estimated from
logit(PoC)=a0 + a1*f + a2*1[obs==1] This seems inverted from the PoP
defintion from Will Kleiber's code, which reads to me like it should be 1-PoP
}
