% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ts_forecast.R
\name{get_quantile_reliability}
\alias{get_quantile_reliability}
\title{Evaluate forecast reliability by evaluating the actual hit rate of the
quantile bins}
\usage{
get_quantile_reliability(ts, tel, quantiles = NA, ...)
}
\arguments{
\item{ts}{A ts_forecast object}

\item{tel}{A list of the telemetry values}

\item{quantiles}{A list of the quantiles to evaluate}

\item{...}{optional arguments to equalize_telemetry_forecast_length}
}
\value{
list of the quantiles and their hit rates
}
\description{
Evaluate forecast reliability by evaluating the actual hit rate of the
quantile bins
}
