% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ts_forecast.R
\name{get_quantile_time_series}
\alias{get_quantile_time_series}
\title{Get a time-series of the forecast value at a particular quantile}
\usage{
get_quantile_time_series(x, alpha)
}
\arguments{
\item{x}{A ts_forecast object}

\item{alpha}{Quantile of interest, numeric [0, 100]}
}
\value{
Numeric vector
}
\description{
Get a time-series of the forecast value at a particular quantile
}
