% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bma.R
\name{get_rho}
\alias{get_rho}
\title{Get rho parameter of beta distribution for single instance}
\usage{
get_rho(FCST, B0, B1, B_transform = NA)
}
\arguments{
\item{FCST}{single forecast}

\item{B0}{intercept of linear model of mean}

\item{B1}{slope of linear model of mean}

\item{B_tranform}{Function of forecast transformation for B coefficients
(optional), e.g. function(x) return(log(x)+1)}
}
\description{
Get rho parameter of beta distribution for single instance
}
