% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prob_forecast.R
\name{get_samples.prob_nd_empirical_forecast}
\alias{get_samples.prob_nd_empirical_forecast}
\title{Sample the empirical copula model and sum to calculate samples of the univariate, aggregate power forecast}
\usage{
\method{get_samples}{prob_nd_empirical_forecast}(x)
}
\value{
A column matrix of aggregate powers
}
\description{
Sample the empirical copula model and sum to calculate samples of the univariate, aggregate power forecast
}
