% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prob_forecast.R
\name{get_shape_params}
\alias{get_shape_params}
\title{Get shape parameters for either a beta or a truncated normal distribution}
\usage{
get_shape_params(x)
}
\arguments{
\item{x}{A fc_bma object}
}
\value{
a list of param1s (alpha for a beta distribution; mean for a truncnorm), param2s (beta for a beta distribution; sd for a truncnorm), and POC (probability of clipping)
}
\description{
Get shape parameters for either a beta or a truncated normal distribution
}
