% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bma.R
\name{get_sigma}
\alias{get_sigma}
\title{Get sigma parameter of distribution for single instance}
\usage{
get_sigma(mu, C0)
}
\arguments{
\item{mu}{mean value for this forecast (AKA rho)}

\item{C0}{height parameter of quadratic model of variance}
}
\description{
Get sigma parameter of distribution for single instance
}
