% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ts_forecast.R
\name{get_sundown_and_NaN_stats}
\alias{get_sundown_and_NaN_stats}
\title{Get sun-up, sun-down, and missing data statistics}
\usage{
get_sundown_and_NaN_stats(ts, tel, ...)
}
\arguments{
\item{tel}{A vector of the telemetry values}

\item{...}{optional arguments to equalize_telemetry_forecast_length}

\item{len_ts}{Number of timesteps in the time series forecast}

\item{return}{A list of summary statistics}
}
\description{
Get statistics on the prevalence of NaN's in telemetry and sun-up/sun-down
discrepancies
}
