% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bma.R
\name{get_weighted_probability}
\alias{get_weighted_probability}
\title{Get z=weighted probability for single instance Returns NA for missing values
and observations exactly at 0 Defines clipping with a percentage threshold,
lambda density is (PoC/(1-lambda))*1[obs>=lambda] +
((1-PoC)/CDF(lambda))*Beta(obs,a,b)*1[obs < lambda]}
\usage{
get_weighted_probability(OBS, PoC, param1, param2, w,
  percent_clipping_threshold, bma_distribution, max_power)
}
\arguments{
\item{OBS}{A single observation}

\item{PoC}{Probability of clipping}

\item{param1}{Distribution parameter 1: mean for both beta and truncnorm
distributions}

\item{param2}{Distribution parameter 2: gamma for a beta; sd for a truncnorm
distribution}

\item{w}{this member's weight}

\item{percent_clipping_threshold}{[0,..1]}

\item{bma_distribution}{"beta" or "truncnorm"}
}
\description{
Get z=weighted probability for single instance Returns NA for missing values
and observations exactly at 0 Defines clipping with a percentage threshold,
lambda density is (PoC/(1-lambda))*1[obs>=lambda] +
((1-PoC)/CDF(lambda))*Beta(obs,a,b)*1[obs < lambda]
}
