% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prob_forecast.R
\name{is.prob_forecast}
\alias{is.prob_forecast}
\title{Check probabilistic forecast class}
\usage{
is.prob_forecast(x)
}
\arguments{
\item{x}{A prob_forecast object}
}
\description{
Check probabilistic forecast class
}
