% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prob_forecast.R
\name{IS}
\alias{IS}
\title{Calculate interval score}
\usage{
IS(x, tel, alpha)
}
\arguments{
\item{x}{A prob_forecast object}

\item{tel}{The realized value}

\item{alpha}{Numeric, to identify the (1-alpha)*100\% quantile of interest}
}
\description{
Calculate interval score for an interval from alpha/2 to 1-alpha/2.
Negatively oriented First term: sharpness, second and third terms: penalty
for reliability.
}
