% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ts_forecast.R
\name{IS_avg}
\alias{IS_avg}
\title{Calculate average interval score}
\usage{
IS_avg(ts, tel, alpha, normalize.by = 1, ...)
}
\arguments{
\item{ts}{A ts_forecast object}

\item{tel}{A list of the telemetry values}

\item{alpha}{Numeric, to identify the (1-alpha)*100\% quantile of interest}

\item{normalize.by}{(optional) A normalization factor, either a scalar or
vector}

\item{...}{additional optional arguments to
equalize_telemetry_forecast_length}
}
\value{
the average IS value
}
\description{
Get average interval score, for an interval from alpha/2 to 1-alpha/2.
Negatively oriented (smaller is better) Characterizes sharpness, with a
penalty for reliability
}
