% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prob_forecast.R
\name{length.prob_forecast}
\alias{length.prob_forecast}
\title{Calculate number of sites in this spatial aggregate}
\usage{
\method{length}{prob_forecast}(x)
}
\arguments{
\item{x}{A prob_forecast object}
}
\value{
Number of sites
}
\description{
Calculate number of sites in this spatial aggregate
}
