% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ts_forecast.R
\name{length.ts_forecast}
\alias{length.ts_forecast}
\title{Calculate number of steps in time series forecast}
\usage{
\method{length}{ts_forecast}(x)
}
\description{
Calculate number of steps in time series forecast
}
