% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dhhh4sims.R
\name{logs_hhh4sims}
\alias{logs_hhh4sims}
\title{Simulation-Based Logarithmic Score Using \code{dhhh4sims}}
\usage{
logs_hhh4sims(observed = NULL, sims, model)
}
\arguments{
\item{observed}{a vector or matrix of observed counts during the
simulation period. By default (\code{NULL}), this is taken from
\code{attr(sims, "stsObserved")}.}

\item{sims}{a \code{"hhh4sims"} object from
\code{\link[surveillance:hhh4_simulate]{surveillance::simulate.hhh4()}}.}

\item{model}{the \code{\link[surveillance:hhh4]{surveillance::hhh4()}} fit underlying \code{sims}.}
}
\value{
a vector or matrix of log-scores for the \code{observed} counts.
}
\description{
The function \code{logs_hhh4sims} computes the logarithmic score of the
forecast distributions based on a \code{\link[surveillance:hhh4]{surveillance::hhh4()}} \code{model} and
\link[surveillance:hhh4_simulate]{simulations} (\code{sims}) thereof. The
forecast distributions are obtained via \code{\link[=dhhh4sims]{dhhh4sims()}} as sequential
mixtures of negative binomial (or Poisson) distributions, which is
different from the kernel density estimation approach employed in
\code{\link[=scores_sample]{scores_sample()}}.
}
\seealso{
\code{\link[=scores_sample]{scores_sample()}} for an alternative approach of calculating
the logarithmic score from simulation-based forecasts
}
\author{
Sebastian Meyer
}
\keyword{univar}
