% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dnbmix.R
\name{logs_nbmix}
\alias{logs_nbmix}
\title{Simulation-Based Logarithmic Score Via \code{dnbmix}}
\usage{
logs_nbmix(observed, means, size)
}
\arguments{
\item{observed}{a vector of observed counts during the simulation
period.}

\item{means}{a \code{n.ahead} x \code{n.sim} matrix of means.}

\item{size}{the dispersion parameter of the \code{\link[=dnbinom]{dnbinom()}} distribution
or \code{NULL} (Poisson forecasts). Can also be time-varying (of length
\code{n.ahead}).}
}
\value{
a vector of log-scores for the \code{observed} counts.
}
\description{
The function \code{logs_nbmix} computes the logarithmic score of forecasts
based on mixtures of negative binomial (or Poisson) distributions via
\code{\link[=dnbmix]{dnbmix()}}. This is different from the kernel density estimation
approach available via \code{\link[=scores_sample]{scores_sample()}}.
}
\seealso{
\code{\link[=scores_sample]{scores_sample()}} for an alternative approach of calculating
the logarithmic score from simulation-based forecasts
}
\author{
Sebastian Meyer
}
\keyword{univar}
