% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ts_forecast.R
\name{MAE}
\alias{MAE}
\title{Get mean absolute error between the forecast median and the actual value}
\usage{
MAE(ts, tel, normalize.by = 1, ...)
}
\arguments{
\item{ts}{A ts_forecast object}

\item{tel}{A list of the telemetry values}

\item{normalize.by}{(optional) A normalization factor, either a scalar or
vector}

\item{...}{optional arguments to equalize_telemetry_forecast_length}
}
\value{
the MAE value
}
\description{
Get mean absolute error between the forecast median and the actual value
}
