% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/marg_transform.R
\name{marg_transform}
\alias{marg_transform}
\title{A structure for transforming back and forth to uniform domain}
\usage{
marg_transform(x, cdf.method = "geenens", ...)
}
\arguments{
\item{x}{A vector of samples or a pre-trained prob_forecast object}

\item{cdf.method}{Method for estimating the marginal distribution. One of 'empirical', 'kde1d', 'geenens', 'kernsmooth', "precalcbma"}

\item{...}{Optional arguments to the distribution estimator. 'xmax' can be used for 'kde1d' and 'geenens'}
}
\description{
A structure for transforming back and forth to uniform domain
}
