% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/marginal_methods.R
\name{marginal_lookup}
\alias{marginal_lookup}
\title{Look-up function for the methods implemented here}
\usage{
marginal_lookup(method)
}
\arguments{
\item{method}{Identifying name: one of 'empirical', 'kde1d', 'geenens', 'kernsmooth', 'precalcbma"}
}
\value{
A function
}
\description{
Look-up function for the methods implemented here
}
