% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/osaplot.R
\name{osaplot}
\alias{osaplot}
\title{Plot One-Step-Ahead Forecasts with Scores}
\usage{
osaplot(quantiles, probs, observed, scores, start = 1, xlab = "Time",
  fan.args = list(), observed.args = list(), key.args = list(), ...,
  heights = c(2, 1))
}
\arguments{
\item{quantiles}{a time x \code{probs} matrix of forecast quantiles at each time
point.}

\item{probs}{numeric vector of probabilities with values between 0 and 1.}

\item{observed}{numeric vector of observed values. Can be \code{NULL}.}

\item{scores}{numeric vector (or matrix) of associated scores.}

\item{start}{time index (x-coordinate) of the first prediction.}

\item{xlab}{x-axis label.}

\item{fan.args}{a list of graphical parameters for the \code{\link[fanplot:fan]{fanplot::fan()}},
e.g., to employ a different \code{\link[grDevices:colorRampPalette]{grDevices::colorRampPalette()}} as
\code{fan.col}, or to enable contour lines via \code{ln}.}

\item{observed.args}{a list of graphical parameters for \code{\link[graphics:lines]{graphics::lines()}}
to modify the plotting style of the \code{observed} values.}

\item{key.args}{if a list, a color key (in \code{\link[fanplot:fan]{fanplot::fan()}}'s
\code{"boxfan"}-style) is added to the fan chart. The list may include
positioning parameters \code{start} (the x-position) and \code{ylim} (the y-range
of the color key), \code{space} to modify the width of the boxfan, and \code{rlab}
to modify the labels. An alternative way of labeling the quantiles is
via the argument \code{ln} in \code{fan.args}.}

\item{...}{further arguments are passed to \code{\link[graphics:plot.default]{graphics::plot.default()}}.}

\item{heights}{numeric vector of length 2 specifying the relative height of
the two subplots.}
}
\description{
This function produces a fan chart of sequential one-step-ahead forecasts.
Dots are added for the observed values. A \code{\link[graphics:matplot]{graphics::matplot()}} of score
values at each time point is added below ("slicing").
}
\author{
Sebastian Meyer
}
