% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prob_forecast.R
\name{pdf_subfunction}
\alias{pdf_subfunction}
\title{Get probability density for individual ensemble member
Includes continuous estimation of discrete component, starting at the threshold and extending to 1}
\usage{
pdf_subfunction(param1, param2, poc, w, xseq, i.thresh, discrete,
  bma_distribution, max_power)
}
\arguments{
\item{param1}{First parameter: alpha for a beta distribution; mean for a truncnorm}

\item{param2}{First parameter: beta for a beta distribution; sd for a truncnorm}

\item{bma_distribution}{"beta" or "truncnorm"}
}
\description{
Get probability density for individual ensemble member
Includes continuous estimation of discrete component, starting at the threshold and extending to 1
}
