% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/marg_transform.R
\name{plot.marg_transform}
\alias{plot.marg_transform}
\title{Plot a marginal transform}
\usage{
\method{plot}{marg_transform}(c)
}
\arguments{
\item{c}{A marg_transform object}
}
\description{
Plot a marginal transform
}
