% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prob_forecast.R
\name{plot.prob_forecast}
\alias{plot.prob_forecast}
\title{Plot probabilistic forecast's estimated pdf (with kde)
Note that CVaR and VaR, while represented on the graph, are calculated directly from sampled data rather than estimated
from the kde results.}
\usage{
\method{plot}{prob_forecast}(x, ...)
}
\description{
Plot probabilistic forecast's estimated pdf (with kde)
Note that CVaR and VaR, while represented on the graph, are calculated directly from sampled data rather than estimated
from the kde results.
}
