% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ts_forecast.R
\name{plot_brier_by_power}
\alias{plot_brier_by_power}
\title{Plot Brier score in terms of power}
\usage{
plot_brier_by_power(ts, tel, xseq)
}
\arguments{
\item{ts}{A ts_forecast object}

\item{tel}{A vector of the telemetry values}

\item{xseq}{A vector of the power thresholds to use}
}
\description{
Plot Brier score in terms of power
}
