% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ts_forecast.R
\name{plot_brier_by_quantile}
\alias{plot_brier_by_quantile}
\title{Plot Brier score along the quantiles from 1 to 99}
\usage{
plot_brier_by_quantile(ts, tel, nmem = NA)
}
\arguments{
\item{ts}{A ts_forecast object}

\item{tel}{A vector of the telemetry values}

\item{nmem}{Number of ensemble members, to illustrate the n+1 bins}
}
\description{
Plot Brier score along the quantiles from 1 to 99
}
