% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prob_forecast.R
\name{plot_crps}
\alias{plot_crps}
\title{Plot probabilistic forecast's CRPS illustration}
\usage{
plot_crps(x, tel)
}
\arguments{
\item{x}{A prob_forecast object}

\item{tel}{telemetry}
}
\description{
Plot probabilistic forecast's CRPS illustration
}
