% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ts_forecast.R
\name{plot_cvar_over_time}
\alias{plot_cvar_over_time}
\title{Plot a time-series of the upper and lower tail CVAR's}
\usage{
plot_cvar_over_time(x)
}
\arguments{
\item{x}{A ts_forecast object}
}
\description{
Plot a time-series of the upper and lower tail CVAR's
}
