% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ts_forecast.R
\name{plot_interval_width}
\alias{plot_interval_width}
\title{Plot average interval width at different nominal intervals}
\usage{
plot_interval_width(ts, tel, alphas = seq(0.9, 0.1, by = -0.1),
  normalize.by = 1, fname = NULL)
}
\arguments{
\item{ts}{A ts_forecast object}

\item{tel}{A list of the telemetry values}

\item{alphas}{(optional) Vector of central (1-alpha)*100\% intervals to calculate}

\item{normalize.by}{(optional) Value or vector to normalize by}

\item{fname}{(optional) file name to save plot}
}
\description{
Plot average interval width at different nominal intervals
}
\seealso{
Other plots: \code{\link{plot_PIT_histogram}},
  \code{\link{plot_quantile_score}},
  \code{\link{plot_reliability}}
}
\concept{plots}
