% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prob_forecast.R
\name{plot_pdf.fc_bma}
\alias{plot_pdf.fc_bma}
\title{Plot APPROXIMATION of the BMA probability density function, including the member component contributributions}
\usage{
\method{plot_pdf}{fc_bma}(x, actual = NA, ymax = NA, normalize = F,
  discrete = T)
}
\arguments{
\item{discrete}{Boolean on whether to plot a discrete component as approximation or the continuous equivalent above the clipping threshold}
}
\description{
Plot APPROXIMATION of the BMA probability density function, including the member component contributributions
}
