% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prob_forecast.R
\name{plot_pdf.fc_emos}
\alias{plot_pdf.fc_emos}
\title{Plot EMOS forecast}
\usage{
\method{plot_pdf}{fc_emos}(x, actual = NA, ymax = NA, normalize = F)
}
\arguments{
\item{x}{A fc_emos object}

\item{actual}{telemetry value}

\item{ymax}{Maximum value to use on the y axis}

\item{normalize}{Boolean of whether to normalize to site's maximum power}
}
\description{
Plot EMOS forecast
}
