% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/marg_transform.R
\name{plot_pdf.marg_transform}
\alias{plot_pdf.marg_transform}
\title{Plot a marg_transform PDF}
\usage{
\method{plot_pdf}{marg_transform}(c, col = "black")
}
\arguments{
\item{c}{A marg_transform object}
}
\description{
Plot a marg_transform PDF
}
