% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ts_forecast.R
\name{plot_PIT_histogram}
\alias{plot_PIT_histogram}
\title{Plot probability integral transform histogram}
\usage{
plot_PIT_histogram(ts, tel, nbins = 10)
}
\arguments{
\item{ts}{A ts_forecast object}

\item{tel}{A list of the telemetry values}

\item{nbins}{To specify number of bars in the histogram}
}
\description{
Plot probability integral transform histogram
}
