% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ts_forecast.R
\name{plot_quantile_score}
\alias{plot_quantile_score}
\title{Plot quantile score vs. quantiles}
\usage{
plot_quantile_score(ts, tel, quantiles = seq(0.01, 0.99, by = 0.01),
  weighting = "none")
}
\arguments{
\item{ts}{A ts_forecast object}

\item{tel}{A list of the telemetry values}

\item{quantiles}{A vector in (0,1)}

\item{weighting}{One of "none" (default), "tails", "right", "left", "center"}
}
\value{
ggplot2 plot
}
\description{
Plot quantile score vs. quantiles
}
