% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ts_forecast.R
\name{plot_reliability}
\alias{plot_reliability}
\title{Plot reliability}
\usage{
plot_reliability(ts, tel, ...)
}
\arguments{
\item{ts}{A ts_forecast object}

\item{tel}{A list of the telemetry values}

\item{...}{optional arguments to equalize_telemetry_forecast_length}
}
\description{
Plot diagonal line diagram of quantiles + observations (i.e., a P-P plot)
}
