% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ts_forecast.R
\name{PRI}
\alias{PRI}
\title{Get the percentile reliability index (PRI) of quantiles 1 to 99. An extension
of the Delle Monache 2006 to look at distance from probabilistic calibration
(P-P plot). * Note that this looks at the performance at the quantiles, not
for each bin -- that is, it is assessed over 99 rather than 100 values}
\usage{
PRI(ts, tel, ...)
}
\arguments{
\item{ts}{A ts_forecast object}

\item{tel}{A vector of the telemetry values}

\item{...}{optional arguments to equalize_telemetry_forecast_length}
}
\description{
Get the percentile reliability index (PRI) of quantiles 1 to 99. An extension
of the Delle Monache 2006 to look at distance from probabilistic calibration
(P-P plot). * Note that this looks at the performance at the quantiles, not
for each bin -- that is, it is assessed over 99 rather than 100 values
}
