% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prob_forecast.R
\name{prob_1d_site_forecast}
\alias{prob_1d_site_forecast}
\title{Initialize a univariate probabilistic power forecast for a specific time point.
Assumes training data already captures differences in magnitude (i.e., power rating) amongst sites.}
\usage{
prob_1d_site_forecast(dat, location, time, n = 3000, epsilon = c(0.05,
  0.95))
}
\arguments{
\item{dat}{A matrix of training data, [ntrain x 1]}

\item{location}{A string}

\item{time}{A lubridate time stamp}

\item{n}{An integer, number of copula samples to take}

\item{epsilon}{Probability levels for lower/upper tail VaR/CVaR calculations, defaults to c(0.05, 0.95)}
}
\value{
An n-dimensional probabilistic forecast object from vine copulas
}
\description{
Initialize a univariate probabilistic power forecast for a specific time point.
Assumes training data already captures differences in magnitude (i.e., power rating) amongst sites.
}
