% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/marginal_methods.R
\name{probempirical}
\alias{probempirical}
\title{Get probability using empirical CDF, with a default minimum x value of 0}
\usage{
probempirical(x, xmax = max(x, na.rm = T))
}
\arguments{
\item{x}{A vector of samples}

\item{xmax}{maximum}
}
\value{
A list of the evaluation points, density, and cumulative distribution
}
\description{
Get probability using empirical CDF, with a default minimum x value of 0
}
