% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/marginal_methods.R
\name{probkde}
\alias{probkde}
\title{Get continuous KDE using KernSmooth based on https://vita.had.co.nz/papers/density-estimation.pdf}
\usage{
probkde(x, ...)
}
\arguments{
\item{x}{A vector of samples}

\item{...}{Optional inputs to the bkde function}
}
\value{
A list of the evaluation points, density, and cumulative distribution
}
\description{
Get continuous KDE using KernSmooth based on https://vita.had.co.nz/papers/density-estimation.pdf
}
