% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/marginal_methods.R
\name{probkde1d}
\alias{probkde1d}
\title{Get KDE using kde1d}
\usage{
probkde1d(x, ...)
}
\arguments{
\item{x}{A vector of samples}

\item{...}{Optional inputs to the bkde function}
}
\value{
A list of the evaluation points, density, and cumulative distribution
}
\description{
Get KDE using kde1d
}
