% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/marginal_methods.R
\name{probprecalc}
\alias{probprecalc}
\title{Reformat from an existing prob_1d_bma_forecast object}
\usage{
probprecalc(x, ...)
}
\arguments{
\item{x}{A vector of samples}

\item{...}{Optional inputs to the bkde function}
}
\value{
A list of the evaluation points, density, and cumulative distribution
}
\description{
Reformat from an existing prob_1d_bma_forecast object
}
