% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/marginal_methods.R
\name{probtranskde}
\alias{probtranskde}
\title{Get KDE using Geenens et. al 2014 and 2018's methods: 1) transformation with probit or log function,
2) local likelihood estimation
3) nearest-neighbor bandwdith selection}
\usage{
probtranskde(x, xmax, scale = 0.9999, zero_offset = 1e-04,
  max_scaler = 2, weight = "WLSCV2", n.res = 500)
}
\arguments{
\item{x}{A vector of samples}

\item{xmax}{Maximum allowable x value for probit transformation, or NaN for log transformation (non-negative)}

\item{scale}{Scaling factor (0,1] to move maximum values off boundary at 1}

\item{zero_offset}{Amount to shift minimum values off the boundary at 0 in the (0,1) domain}

\item{max_scaler}{For log transforms, estimation is made over the range max(x)*max_scaler}

\item{weight}{One of 'LSCV', 'WLSCV1', 'WLSCV2'}

\item{n.res}{length of resulting estimation vector}
}
\value{
A list of the estimate evaluation points, density, and cumulative distribution
}
\description{
Get KDE using Geenens et. al 2014 and 2018's methods: 1) transformation with probit or log function,
2) local likelihood estimation
3) nearest-neighbor bandwdith selection
}
