% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prob_forecast.R
\name{qc_bma_input}
\alias{qc_bma_input}
\title{Specific quality control handling for BMA forecasts, for missing members that have already been trained in the model.
Reallocates member weights if necessary}
\usage{
qc_bma_input(members, model)
}
\description{
Specific quality control handling for BMA forecasts, for missing members that have already been trained in the model.
Reallocates member weights if necessary
}
