% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prob_forecast.R
\name{qc_input}
\alias{qc_input}
\title{Quality control 1-dimension input data}
\usage{
qc_input(dat)
}
\arguments{
\item{dat}{A numeric vector of data}
}
\value{
data without NA's
}
\description{
Quality control 1-dimension input data
}
