% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ts_forecast.R
\name{QS}
\alias{QS}
\title{Get weighted quantile score at certain quantile(s)}
\usage{
QS(ts, tel, quantiles)
}
\arguments{
\item{ts}{A ts_forecast object}

\item{tel}{A list of the telemetry values}

\item{quantiles, }{numeric [0,1].  Can be a scalar or a vector}
}
\value{
the Quantile score
}
\description{
Get weighted quantile score at certain quantile(s)
}
