% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ts_forecast.R
\name{RI}
\alias{RI}
\title{Get the reliability index (RI) as per Delle Monache 2006}
\usage{
RI(ts, tel, nbins = 100, ...)
}
\arguments{
\item{ts}{A ts_forecast object}

\item{tel}{A vector of the telemetry values}

\item{nbins}{Number of bins (original ensemble members) to evaluate}

\item{...}{optional arguments to equalize_telemetry_forecast_length}
}
\description{
Get the reliability index (RI) as per Delle Monache 2006
}
