% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/marginal_methods.R
\name{scale_01}
\alias{scale_01}
\title{Scale from [0, xmax] to (0,1). Values exactly at 0 are shifted up by an offset to stabilize the boundary.
All other values are scaled to avoid the boundary at 1.}
\usage{
scale_01(x, xmax, scale, zero_offset)
}
\arguments{
\item{x}{A vector of samples}

\item{xmax}{Maximum allowable x value. If NA, function returns the input values.}

\item{scale}{Scaling factor (0,1] to move maximum values off boundary at 1}

\item{zero_offset}{Amount to shift minimum values off the boundary at 0 in the (0,1) domain}
}
\value{
The vector of samples rescaled to [0,1]
}
\description{
Scale from [0, xmax] to (0,1). Values exactly at 0 are shifted up by an offset to stabilize the boundary.
All other values are scaled to avoid the boundary at 1.
}
