% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/marginal_methods.R
\name{scale_full}
\alias{scale_full}
\title{Scale from (0,1) to [0, xmax]. The slight offset at 0 is not recovered here.}
\usage{
scale_full(x01, d01, xmax)
}
\arguments{
\item{x01}{A vector of points in (0,1)}

\item{d01}{Associated density}

\item{xmax}{Maximum allowable x value. If NA, function returns the input values.}
}
\value{
The vector of samples rescaled to [0,1]
}
\description{
Scale from (0,1) to [0, xmax]. The slight offset at 0 is not recovered here.
}
