% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scores.R
\name{scores_lnorm}
\alias{scores_lnorm}
\alias{dss_lnorm}
\alias{logs_lnorm}
\title{Proper Scoring Rules for Log-Normal Forecasts}
\usage{
dss_lnorm(x, meanlog, sdlog)

logs_lnorm(x, meanlog, sdlog)
}
\arguments{
\item{x}{the observed counts.}

\item{meanlog, sdlog}{parameters of the log-normal distribution, i.e., mean
and standard deviation of the distribution on the log scale.}
}
\value{
scores for the predictions of the observations in \code{x} (maintaining
their dimensions).
}
\description{
Predictions are assumed to have a \link[=plnorm]{LN}(\code{meanlog}, \code{sdlog})
distribution. The functions are vectorized and preserve the dimensions of
the input.
}
\section{Functions}{
\itemize{
\item \code{dss_lnorm}: Computes the Dawid-Sebastiani score (DSS) for the
log-normal distribution similar to \code{\link[surveillance:dss]{surveillance::dss()}} for the Poisson
and negative binomial distribution.

\item \code{logs_lnorm}: Computes the logarithmic score (LogS) for the
log-normal distribution similar to \code{\link[surveillance:logs]{surveillance::logs()}} for the Poisson
and negative binomial distribution.
}}

\seealso{
The R package \CRANpkg{scoringRules} implements the logarithmic
score and the continuous ranked probability scores (CRPS) for many
continuous distributions.
}
\author{
Sebastian Meyer
}
