% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scores_lnorm_discrete.R
\name{scores_lnorm_discrete}
\alias{scores_lnorm_discrete}
\title{Proper Scoring Rules for \emph{Discretized} Log-Normal Forecasts}
\usage{
scores_lnorm_discrete(x, meanlog, sdlog, which = c("dss", "logs"))
}
\arguments{
\item{x}{the observed counts.}

\item{meanlog}{parameters of the log-normal distribution, i.e., mean
and standard deviation of the distribution on the log scale.}

\item{sdlog}{parameters of the log-normal distribution, i.e., mean
and standard deviation of the distribution on the log scale.}

\item{which}{a character vector specifying which scoring rules to apply. The
Dawid-Sebastiani score (\code{"dss"}) and the logarithmic score (\code{"logs"})
are available and both computed by default.}
}
\value{
scores for the predictions of the observations in \code{x} (maintaining
their dimensions).
}
\description{
Compute scores for discretized log-normal forecasts.
The function is vectorized and preserves the dimension of the input.
}
