% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scores_sample.R
\name{scores_sample}
\alias{scores_sample}
\title{Proper Scoring Rules based on Simulations}
\usage{
scores_sample(x, sims, which = c("dss", "logs"))
}
\arguments{
\item{x}{a vector of observed counts.}

\item{sims}{a matrix of simulated counts with as many rows as \code{length(x)}.}

\item{which}{a character vector specifying which scoring rules to apply. The
Dawid-Sebastiani score (\code{"dss"}) and the logarithmic score (\code{"logs"})
are available and both computed by default.}
}
\value{
scores for the predictions of the observations in \code{x} (maintaining
their dimensions).
}
\description{
This is a simple wrapper around functions from the \CRANpkg{scoringRules}
package to calculate scoring rules from simulation-based forecasts.
Calculation of the logarithmic score involves kernel density estimation,
see \code{\link[scoringRules:scores_sample_univ]{scoringRules::logs_sample()}}.
The function is vectorized and preserves the dimension of the input.
}
