% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prob_forecast.R
\name{sharpness}
\alias{sharpness}
\title{Calculate sharpness for an interval from alpha/2 to 1-alpha/2. Negatively
oriented}
\usage{
sharpness(x, tel, alpha)
}
\arguments{
\item{x}{A prob_forecast object}

\item{tel}{(unneeded) For calling signature consistency with other metrics}

\item{alpha}{Numeric, to identify the (1-alpha)*100\% quantile of interest}
}
\description{
Calculate sharpness for an interval from alpha/2 to 1-alpha/2. Negatively
oriented
}
