% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ts_forecast.R
\name{sharpness_avg}
\alias{sharpness_avg}
\title{Average interval width}
\usage{
sharpness_avg(ts, tel, alpha, normalize.by = 1, ...)
}
\arguments{
\item{ts}{A ts_forecast object}

\item{tel}{A vector of telemetry values.}

\item{alpha}{Numeric, to identify the (1-alpha)*100\% quantile of interest}

\item{normalize.by}{(optional) A normalization factor, either a scalar or
vector}

\item{...}{additional optional arguments to
equalize_telemetry_forecast_length}
}
\value{
the average sharpness value (interval width)
}
\description{
Get average sharpness (interval width), for an interval from alpha/2 to
1-alpha/2. Negatively oriented (smaller is better) Telemetry isn't used in
sharpness, but it is only averaged over times when telemetry is available for
direct compatibility with other metrics.
}
