% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/emos.R
\name{tnorm_crps_obj}
\alias{tnorm_crps_obj}
\title{Subfunction to get average crps over the training period}
\usage{
tnorm_crps_obj(par, obs, ens, max_power)
}
\arguments{
\item{par}{a vector of parameters to optimize over: c, d, a, b1, ... bn}

\item{obs}{a vector of observations of length m}

\item{ens}{a [time x member] (i.e., [m x n]) matrix of the ensemble forecasts
for each observation}

\item{max_power}{Site AC power, for upper limit of truncated normal}
}
\description{
Subfunction to get average crps over the training period
}
