% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/marg_transform.R
\name{to_probability}
\alias{to_probability}
\title{Transform from variable to probability density}
\usage{
to_probability(c, x)
}
\arguments{
\item{c}{A marg_transform object}

\item{x}{A vector of evaluation points to transform}
}
\value{
A vector linearly interpolated from the variable to probability density
}
\description{
Transform from variable to probability density
}
