% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/marg_transform.R
\name{to_uniform}
\alias{to_uniform}
\title{Transform from variable to uniform domain}
\usage{
to_uniform(c, x)
}
\arguments{
\item{c}{A marg_transform object}

\item{x}{A vector of evaluation points to transform}
}
\value{
A vector linearly interpolated from the variable to uniform domain
}
\description{
Transform from variable to uniform domain
}
