% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/marginal_methods.R
\name{transform_to_real_line}
\alias{transform_to_real_line}
\title{Transform values from bounded to unbounded support through probit or log transform. Assumes lower boundary is at 0.
Author: Gery Geenens and Kate Doubleday}
\usage{
transform_to_real_line(x, xmax, scale, zero_offset)
}
\arguments{
\item{x}{A vector of samples}

\item{xmax}{Maximum allowable x value.}

\item{scale}{Scaling factor (0,1] to move maximum values off boundary at 1}

\item{zero_offset}{Amount to shift minimum values off the boundary at 0 in the (0,1) domain}
}
\value{
The vector of transformed samples
}
\description{
Transform values from bounded to unbounded support through probit or log transform. Assumes lower boundary is at 0.
Author: Gery Geenens and Kate Doubleday
}
