% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ts_forecast.R
\name{ts_forecast.list}
\alias{ts_forecast.list}
\title{Calculate a timeseries forecast from a list}
\usage{
\method{ts_forecast}{list}(x, start_time, time_step, scale, location,
  method)
}
\arguments{
\item{x}{A list of prob_forecast objects}

\item{start_time}{A lubridate time stamp}

\item{time_step}{Time step in hours}

\item{scale}{One of 'site', 'region', 'total'}

\item{location}{A string}

\item{method}{A string}
}
\description{
An alternative ts_forecast constructor, which accepts an already calculated
list of forecasts. For future work, this could be integrated with copula
post-processing with site-level forecasts.
}
