% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ts_forecast.R
\name{ts_forecast.matrix}
\alias{ts_forecast.matrix}
\title{Calculate a timeseries forecast from a matrix}
\usage{
\method{ts_forecast}{matrix}(x, start_time, time_step, scale, location,
  method, sun_up_threshold = 0.5, MoreTSArgs = NA, ...)
}
\arguments{
\item{x}{A matrix of training data of dimensions [time x ntrain] for a
1-dimensional forecast}

\item{start_time}{A lubridate time stamp}

\item{time_step}{Time step in hours}

\item{scale}{Must be "site"}

\item{location}{A string}

\item{method}{One of 'gaussian', 'empirical', 'vine' (irrelevant if scale ==
'site')}

\item{sun_up_threshold}{An absolute [MW] threshold on the ensemble members to
remove dubious sunrise/sunset valud}

\item{MoreTSArgs}{An optional dictionary of time-series arguments to the
forecast calculation}

\item{...}{optional arguments to the prob_forecast object}
}
\description{
Construct a time series of power forecasts, using input training data.
Assumes training data already captures differences in magnitude (i.e., power
rating) amongst sites. Forecast is NA for times when sun is down.
}
