% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ts_forecast.R
\name{ts_forecast}
\alias{ts_forecast}
\title{Register ts_forecast constructor to allow for overloading}
\usage{
ts_forecast(x, ...)
}
\arguments{
\item{x}{A ts_forecast object}
}
\description{
Register ts_forecast constructor to allow for overloading
}
