% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/update.Arima.R
\name{update.Arima}
\alias{update.Arima}
\title{Refit an ARIMA Model on a Subset of the Time Series}
\usage{
\method{update}{Arima}(object, subset, ...)
}
\arguments{
\item{object}{an object of class \code{"Arima"}, e.g., from
\code{\link[forecast:auto.arima]{forecast::auto.arima()}}.}

\item{subset}{an integer vector selecting part of the original time series
(and external regressors).}

\item{...}{further arguments to be passed to \code{\link[=arima]{arima()}}.}
}
\value{
the updated model.
}
\description{
There seems to be no function in package \CRANpkg{forecast} (as of version
8.2) to re-estimate an ARIMA model on a subset of the original time series.
This \code{update} method does exactly that.
}
\author{
Sebastian Meyer
}
