% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zzz.R
\name{use_optipng}
\alias{use_optipng}
\title{Should the Package Vignettes Use \code{optipng}?}
\usage{
use_optipng(level = 2)
}
\arguments{
\item{level}{the optimization level (0-7). The default (2) is a good
compromise between compression and speed.}
}
\description{
If the environment variable \env{OPTIPNG} is set to \code{TRUE} (or
something regarded as such via \code{\link[=as.logical]{as.logical()}}), the png files produced
when building the vignettes will be compressed (losslessly) using the
\command{optipng} program.
}
\keyword{internal}
