% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ts_forecast.R
\name{write.ts_forecast}
\alias{write.ts_forecast}
\title{Write ts_forecast quantiles to .csv}
\usage{
write.ts_forecast(x, file, window = NA, percentiles = 1:99)
}
\arguments{
\item{x}{A ts_forecast object}

\item{file}{A file name}

\item{window}{(optional) A vector of (start index, end index) to plot certain
time window}

\item{percentiles}{(optional -- defaults 1st to 99th percentiles) Select the
quantiles associated with which percentiles will be exported}
}
\description{
Export a csv file of [time x quantiles] Note: this follows the convention of
e.g., write.csv and write.table, but write isn't an exported generic function
}
