% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prob_forecast.R
\name{calc_quantiles}
\alias{calc_quantiles}
\title{Calculate forecast quantiles}
\usage{
calc_quantiles(x, samples, quantile_density = 0.01)
}
\arguments{
\item{x}{A probabilistic forecast}

\item{samples}{A column matrix of summed samples}

\item{quantile_density}{Numeric in (0,1), i.e., 0.1 to calculate quantiles at every 10%}
}
\value{
Estimated quantiles
}
\description{
Calculate forecast quantiles
}
