% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ts_forecast.R
\name{check_sunup}
\alias{check_sunup}
\title{Check GHI/power matrix for any positive numbers}
\usage{
check_sunup(x, sun_up_threshold = 0.5)
}
\arguments{
\item{x}{A matrix}
}
\value{
A boolean
}
\description{
Check GHI/power matrix for any positive numbers
}
