% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/osaplot.R
\name{osaplot}
\alias{osaplot}
\title{Plot (One-Step-Ahead) Forecasts with Scores}
\usage{
osaplot(quantiles, probs, means, observed, scores, start = 1, xlab = "Time",
  fan.args = list(), means.args = list(), observed.args = list(),
  key.args = list(), ..., scores.args = list(), legend.args = list(),
  heights = c(0.6, 0.4))
}
\arguments{
\item{quantiles}{a time x \code{probs} matrix of forecast quantiles at each time
point.}

\item{probs}{numeric vector of probabilities with values between 0 and 1.}

\item{means}{(optional) numeric vector of point forecasts at each time
point.}

\item{observed}{(optional) numeric vector of observed values.}

\item{scores}{(optional) numeric vector (or matrix) of associated scores.}

\item{start}{time index (x-coordinate) of the first prediction.}

\item{xlab}{x-axis label.}

\item{fan.args}{a list of graphical parameters for the \code{\link[fanplot:fan]{fanplot::fan()}},
e.g., to employ a different \code{\link[=colorRampPalette]{colorRampPalette()}} as
\code{fan.col}, or to enable contour lines via \code{ln}.}

\item{means.args}{a list of graphical parameters for \code{\link[=lines]{lines()}}
to modify the plotting style of the point predictions.}

\item{observed.args}{a list of graphical parameters for \code{\link[=lines]{lines()}}
to modify the plotting style of the \code{observed} values.}

\item{key.args}{if a list, a color key (in \code{\link[fanplot:fan]{fanplot::fan()}}'s
\code{"boxfan"}-style) is added to the fan chart. The list may include
positioning parameters \code{start} (the x-position) and \code{ylim} (the y-range
of the color key), \code{space} to modify the width of the color key, and
\code{rlab} to modify the labels. An alternative way of labeling the
quantiles is via the argument \code{ln} in \code{fan.args}.}

\item{...}{further arguments are passed to \code{\link[=plot.default]{plot.default()}}.}

\item{scores.args}{a list of graphical parameters for \code{\link[=matplot]{matplot()}} to modify
the style of the \code{scores} subplot at the bottom.}

\item{legend.args}{if a list (of parameters for \code{\link[=legend]{legend()}}) and
\code{ncol(scores) > 1}, a legend is added to the \code{scores} subplot.}

\item{heights}{numeric vector of length 2 specifying the relative height of
the two subplots.}
}
\description{
This function produces a fan chart of sequential (one-step-ahead) forecasts
with dots for the observed values, using \code{\link[surveillance:fanplot]{surveillance::fanplot()}}, which
itself wraps \code{\link[fanplot:fan]{fanplot::fan()}}. A \code{\link[=matplot]{matplot()}} of score
values at each time point is added below ("slicing").
}
\author{
Sebastian Meyer
}
