% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prob_forecast.R
\name{plot.prob_forecast}
\alias{plot.prob_forecast}
\title{Plot probabilistic forecast's quantiles}
\usage{
\method{plot}{prob_forecast}(x)
}
\arguments{
\item{x}{A prob_forecast object}
}
\description{
Plot probabilistic forecast's quantiles
}
