% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ts_forecast.R
\name{plot.ts_forecast}
\alias{plot.ts_forecast}
\title{Plot fan plot of the time-series probabilistic forecast}
\usage{
\method{plot}{ts_forecast}(x, tel = NA, window = NA,
  normalize.by = 1, xbreaks = 4, ylim = NULL, text.size = 14,
  ylab = ggplot2::element_blank(), xlab = ggplot2::element_blank(),
  xticklab = T, yticklab = T)
}
\arguments{
\item{x}{A ts_forecast object}

\item{tel}{vector of telemetry (optional)}

\item{window}{(optional) A vector of (start index, end index) to plot certain
time window}

\item{normalize.by}{(optional) Plant power to normalize results. Defaults to 1}

\item{xbreaks}{(optional) Determines width between x ticks, defaults to 4}

\item{ylim}{(optional) Upper limit of y-axis}

\item{ylab}{(optional) y-axis label}

\item{xlab}{(optional) x-axis label}

\item{xticklab}{(optional) Boolean Whether to label x-axis ticks. Defaults to True}

\item{yticklab}{(optional) Boolean Whether to label y-axis ticks. Defaults to True}
}
\description{
Plot fan plot of the time-series probabilistic forecast
}
