% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ts_forecast.R
\name{plot.ts_forecast}
\alias{plot.ts_forecast}
\title{Plot fan plot of the time-series probabilistic forecast}
\usage{
\method{plot}{ts_forecast}(x, ..., actuals = NA)
}
\arguments{
\item{x}{A ts_forecast object}

\item{actuals}{list of realized values (optional)}
}
\description{
Plot fan plot of the time-series probabilistic forecast
}
