% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zzz.R
\docType{import}
\name{reexports}
\alias{reexports}
\alias{index}
\alias{autoplot.zoo}
\alias{fortify.zoo}
\title{Objects exported from other packages}
\keyword{internal}
\description{
These objects are imported from other packages. Follow the links
below to see their documentation.

\describe{
  \item{zoo}{\code{\link[zoo]{index}}, \code{\link[zoo]{autoplot.zoo}}, \code{\link[zoo]{fortify.zoo}}}
}}

