% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scores_lnorm.R
\name{scores_lnorm}
\alias{scores_lnorm}
\title{Proper Scoring Rules for Log-Normal Forecasts}
\usage{
scores_lnorm(x, meanlog, sdlog, which = c("dss", "logs"))
}
\arguments{
\item{x}{the observed counts.}

\item{meanlog, sdlog}{parameters of the log-normal distribution, i.e., mean
and standard deviation of the distribution on the log scale.}

\item{which}{a character vector specifying which scoring rules to apply. The
Dawid-Sebastiani score (\code{"dss"}) and the logarithmic score (\code{"logs"})
are available and both computed by default.}
}
\value{
scores for the predictions of the observations in \code{x} (maintaining
their dimensions).
}
\description{
This is a simple wrapper around functions from the \CRANpkg{scoringRules}
package for predictions with a \link[=plnorm]{LN}(\code{meanlog}, \code{sdlog})
distribution. The function is vectorized and preserves the dimension of
the input.
}
