% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ts_forecast.R
\name{ts_forecast}
\alias{ts_forecast}
\title{Initialize a time series of power forecasts. Assumes training data already captures
differences in magnitude (i.e., power rating) amongst sites. Forecast is NA for times when sun is down.}
\usage{
ts_forecast(x, start_time, time_step, scale = "region",
  location = "unknown", method = "vine", n = 3000,
  epsilon = c(0.05, 0.95))
}
\arguments{
\item{x}{A list of training data the length of the time-series. Each element should be [ntrain x 1] matrix of training data (for scale=='site')
or a [ntrain x nsites] matrix for scale=='region' or 'total'}

\item{start_time}{A lubridate time stamp}

\item{time_step}{Time step in hours}

\item{scale}{One of 'site', 'region', 'total'}

\item{location}{A string}

\item{method}{One of 'gaussian', 'empirical', 'vine' (irrelevant if scale == 'site')}

\item{n}{An integer: Number of samples to take}

\item{epsilon}{Probability levels for lower/upper tail VaR/CVaR calculations, defaults to c(0.05, 0.95)}
}
\description{
Initialize a time series of power forecasts. Assumes training data already captures
differences in magnitude (i.e., power rating) amongst sites. Forecast is NA for times when sun is down.
}
